+++ setup.cfg
@@ -109,6 +109,7 @@ packages =
     ethereum/prague/vm
     ethereum/prague/vm/instructions
     ethereum/prague/vm/precompiled_contracts
+    ethereum/prague/vm/precompiled_contracts/bls12_381
 
 
 package_dir =
