+++ src/ethereum/prague/vm/precompiled_contracts/bls12_381/__init__.py
@@ -1,6 +1,6 @@
 """
-Ethereum Virtual Machine (EVM) BLS12 381 G2 CONTRACTS
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+BLS12 381 Precompile
+^^^^^^^^^^^^^^^^^^^^
 
 .. contents:: Table of Contents
     :backlinks: none
@@ -9,13 +9,14 @@ Ethereum Virtual Machine (EVM) BLS12 381 G2 CONTRACTS
 Introduction
 ------------
 
-Implementation of pre-compiles in G2 (curve over base prime field).
+Precompile for BLS12-381 curve operations.
 """
 from typing import Tuple
 
 from py_ecc.bls12_381.bls12_381_curve import (
+    FQ,
     FQ2,
-    add,
+    b,
     b2,
     curve_order,
     is_on_curve,
@@ -23,14 +24,14 @@ from py_ecc.bls12_381.bls12_381_curve import (
 )
 from py_ecc.typing import Point2D
 
-from ethereum.base_types import U256, Bytes, Uint
+from ethereum.base_types import U256, Bytes
 
-from ...vm import Evm
-from ...vm.gas import charge_gas
-from ...vm.memory import buffer_read
-from ..exceptions import InvalidParameter
+from ....vm.memory import buffer_read
+from ...exceptions import InvalidParameter
 
-P = FQ2.field_modulus
+P = FQ.field_modulus
+P2 = FQ2.field_modulus
+
 K_DISCOUNT = {
     1: 1200,
     2: 888,
@@ -166,6 +167,100 @@ MAX_DISCOUNT = 174
 MULTIPLIER = 1000
 
 
+def bytes_to_G1(data: Bytes) -> Point2D:
+    """
+    Decode 128 bytes to a G1 point. Does not perform sub-group check.
+
+    Parameters
+    ----------
+    data :
+        The bytes data to decode.
+
+    Returns
+    -------
+    point : Point2D
+        The G1 point.
+
+    Raises
+    ------
+    InvalidParameter
+        Either a field element is invalid or the point is not on the curve.
+    """
+    assert len(data) == 128
+    x = int.from_bytes(data[:64], "big")
+    y = int.from_bytes(data[64:], "big")
+
+    if x >= P:
+        raise InvalidParameter("Invalid field element")
+    if y >= P:
+        raise InvalidParameter("Invalid field element")
+
+    if x == 0 and y == 0:
+        return None
+
+    point = (FQ(x), FQ(y))
+
+    # Check if the point is on the curve
+    if not is_on_curve(point, b):
+        raise InvalidParameter("Point is not on curve")
+
+    return point
+
+
+def G1_to_bytes(point: Point2D) -> Bytes:
+    """
+    Encode a G1 point to 128 bytes.
+
+    Parameters
+    ----------
+    point :
+        The G1 point to encode.
+
+    Returns
+    -------
+    data : Bytes
+        The encoded data.
+    """
+    if point is None:
+        return b"\x00" * 128
+
+    x, y = point
+
+    x_bytes = int(x).to_bytes(64, "big")
+    y_bytes = int(y).to_bytes(64, "big")
+
+    return x_bytes + y_bytes
+
+
+def decode_G1_scalar_pair(data: Bytes) -> Tuple[Point2D, int]:
+    """
+    Decode 160 bytes to a G1 point and a scalar.
+
+    Parameters
+    ----------
+    data :
+        The bytes data to decode.
+
+    Returns
+    -------
+    point : Tuple[Point2D, int]
+        The G1 point and the scalar.
+
+    Raises
+    ------
+    InvalidParameter
+        If the sub-group check failed.
+    """
+    assert len(data) == 160
+    p = bytes_to_G1(buffer_read(data, U256(0), U256(128)))
+    if multiply(p, curve_order) is not None:
+        raise InvalidParameter("Sub-group check failed.")
+
+    m = int.from_bytes(buffer_read(data, U256(128), U256(32)), "big")
+
+    return p, m
+
+
 def bytes_to_FQ2(data: Bytes) -> FQ2:
     """
     Decode 128 bytes to a FQ2 element.
@@ -189,9 +284,9 @@ def bytes_to_FQ2(data: Bytes) -> FQ2:
     c_0 = int.from_bytes(data[:64], "big")
     c_1 = int.from_bytes(data[64:], "big")
 
-    if c_0 >= P:
+    if c_0 >= P2:
         raise InvalidParameter("Invalid field element")
-    if c_1 >= P:
+    if c_1 >= P2:
         raise InvalidParameter("Invalid field element")
 
     return FQ2((c_0, c_1))
@@ -199,7 +294,7 @@ def bytes_to_FQ2(data: Bytes) -> FQ2:
 
 def bytes_to_G2(data: Bytes) -> Point2D:
     """
-    Decode 256 bytes to a G2 point.
+    Decode 256 bytes to a G2 point. Does not perform sub-group check.
 
     Parameters
     ----------
@@ -300,109 +395,3 @@ def decode_G2_scalar_pair(data: Bytes) -> Tuple[Point2D, int]:
     m = int.from_bytes(buffer_read(data, U256(256), U256(32)), "big")
 
     return p, m
-
-
-def bls12_g2_add(evm: Evm) -> None:
-    """
-    The bls12_381 G2 point addition precompile.
-
-    Parameters
-    ----------
-    evm :
-        The current EVM frame.
-
-    Raises
-    ------
-    InvalidParameter
-        If the input length is invalid.
-    """
-    data = evm.message.data
-    if len(data) != 512:
-        raise InvalidParameter("Invalid Input Length")
-
-    # GAS
-    charge_gas(evm, Uint(800))
-
-    # OPERATION
-    p1 = bytes_to_G2(buffer_read(data, U256(0), U256(256)))
-    p2 = bytes_to_G2(buffer_read(data, U256(256), U256(256)))
-
-    result = add(p1, p2)
-
-    evm.output = G2_to_bytes(result)
-
-
-def bls12_g2_multiply(evm: Evm) -> None:
-    """
-    The bls12_381 G2 multiplication precompile.
-
-    Parameters
-    ----------
-    evm :
-        The current EVM frame.
-
-    Raises
-    ------
-    InvalidParameter
-        If the input length is invalid.
-    """
-    data = evm.message.data
-    if len(data) != 288:
-        raise InvalidParameter("Invalid Input Length")
-
-    # GAS
-    charge_gas(evm, Uint(45000))
-
-    # OPERATION
-    p, m = decode_G2_scalar_pair(data)
-    result = multiply(p, m)
-
-    evm.output = G2_to_bytes(result)
-
-
-def bls12_g2_msm(evm: Evm) -> None:
-    """
-    The bls12_381 G2 multi-scalar multiplication precompile.
-    Note: This uses the naive approach to multi-scalar multiplication
-    which is not suitably optimized for production clients. Clients are
-    required to implement a more efficient algorithm such as the Pippenger
-    algorithm.
-
-    Parameters
-    ----------
-    evm :
-        The current EVM frame.
-
-    Raises
-    ------
-    InvalidParameter
-        If the input length is invalid.
-    """
-    data = evm.message.data
-    if len(data) == 0 or len(data) % 288 != 0:
-        raise InvalidParameter("Invalid Input Length")
-
-    # GAS
-    k = len(data) // 288
-    if k <= 128:
-        discount = K_DISCOUNT[k]
-    else:
-        discount = MAX_DISCOUNT
-
-    gas_cost = Uint(k * 45000 * discount // MULTIPLIER)
-    charge_gas(evm, gas_cost)
-
-    # OPERATION
-    for i in range(k):
-        start_index = i * 288
-        end_index = start_index + 288
-
-        p, m = decode_G2_scalar_pair(data[start_index:end_index])
-        product = multiply(p, m)
-
-        if i == 0:
-            result = product
-        else:
-            result = add(result, product)
-
-    evm.output = G2_to_bytes(result)
