+++ src/ethereum/prague/vm/precompiled_contracts/mapping.py
@@ -34,8 +34,16 @@ from . import (
 )
 from .alt_bn128 import alt_bn128_add, alt_bn128_mul, alt_bn128_pairing_check
 from .blake2f import blake2f
-from .bls12_381_g1 import bls12_g1_add, bls12_g1_msm, bls12_g1_multiply
-from .bls12_381_g2 import bls12_g2_add, bls12_g2_msm, bls12_g2_multiply
+from .bls12_381.bls12_381_g1 import (
+    bls12_g1_add,
+    bls12_g1_msm,
+    bls12_g1_multiply,
+)
+from .bls12_381.bls12_381_g2 import (
+    bls12_g2_add,
+    bls12_g2_msm,
+    bls12_g2_multiply,
+)
 from .ecrecover import ecrecover
 from .identity import identity
 from .modexp import modexp
